% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exprs.test}
\alias{exprs.test}
\title{A dataset for testing the function}
\format{
\subsection{\code{exprs.test}}{

A data frame with 156 rows and 40 columns:
\describe{
\item{country}{Country name}
\item{iso2, iso3}{2 & 3 letter ISO country codes}
\item{year}{Year}
...
}
}
}
\source{
AMC
}
\usage{
exprs.test
}
\description{
A dataset for testing the function
}
\keyword{datasets}
