% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NanoCMSer-package.R
\docType{package}
\name{NanoCMSer-package}
\alias{NanoCMSer-package}
\title{NanoCMSer: CMS classifier for NanoString platform}
\description{
NanoCMSer: A Clinically Applicable Stratification Tool for Stratification of Colorectal Cancer Samples
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/atorang/NanoCMSer}
  \item Report bugs at \url{https://github.com/atorang/NanoCMSer/issues}
}

}
\author{
\strong{Maintainer}: Arezo Torang \email{torang.arezo@gmail.com} (\href{https://orcid.org/0000-0002-7912-3964}{ORCID})

}
\keyword{internal}
