% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NanoCMSer.R
\name{NanoCMSer}
\alias{NanoCMSer}
\title{NanoCMSer}
\usage{
NanoCMSer(
  data,
  sample_type = NA,
  perform_log2 = T,
  gene_names = "ensembl",
  impute = T
)
}
\arguments{
\item{data}{a numeric matrix or data.frame with expression levels. Rows are genes and columns are samples.
We recommend using raw count data.}

\item{sample_type}{a character string. Accepted values are: \code{"tumorFF"}; fresh frozen patient samples,
\code{"tumorFFPE"}; formalin-fixed paraffin-embedded patient samples,
\code{"models"}; human in vitro models including cell lines, primary cultures and organoids.}

\item{perform_log2}{if \code{perform_log2 = TRUE}, the data will be log2-transformed.
It is a logical determining if data needs log2-transformation or if the data is already log2-transformed.
Default is \code{perform_log2 = TRUE}.}

\item{gene_names}{a character string specifying the gene annotation used in \code{data}. Accepted values are
\code{"ensembl"}, \code{"symbol"} and \code{"entrez"}. Default value is \code{"ensembl"}.
We advise using Ensembl IDs to mitigate the risk of missing values caused by gene symbol updates.
The current version encompasses all previous versions of gene symbols up to 2024.}

\item{impute}{if \code{impute = TRUE}, for missing genes, up to 10\% of the genes utilized in the classifiers will be imputed
utilizing a trained linear regression model. If \code{impute = FALSE}, the classifier will generate an error message in the event of missing genes.
Default is \code{impute = TRUE}.}
}
\value{
An object of class "\code{data.frame}" containing the classification results.
Columns CMS1-4 show the chance of being each subtype, based on an elastic-net model.
nearestCMS indicates subtype with highest chance, regardless of the cutoff.
predictedCMS, shows subtype of samples that confidently predicted (cuttoff = 0.6).
}
\description{
\code{NanoCMSer} is a classifier developed for the NanoString platform to predict
Consensus Molecular Subtypes (CMS) in colorectal cancer.
}
\examples{
#classify a predefined set
data("exprs.test")
NanoCMSer(data=exprs.test, sample_type="tumorFFPE", perform_log2=TRUE,
gene_names="ensembl",impute=TRUE)
}
